third_party/blink/web_tests/http/tests/xmlhttprequest/response-blob-mimetype-expected.txt

This tests that XMLHttpRequest lowers the MIME Type when creates a response blob.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS returnedMimeType is "multipart/mixed"
PASS returnedMimeType is "text/plain"
PASS successfullyParsed is true

TEST COMPLETE
